fix(console): loop import
