+++ web/console/Wrapper.tsx
@@ -14,6 +14,7 @@ import { Icon, Text, Bubble, NavMenu, List, ExternalLink } from '@tencent/tea-co
 import zhCN from 'antd/lib/locale/zh_CN';
 import { ConfigProvider, Spin } from 'antd';
 import { TkeVersion } from '@/src/modules/common/components/tke-version';
+import { ConsoleModuleEnum } from '@config/platform';
 
 const routerSea = seajs.require('router');
 
@@ -23,36 +24,6 @@ enum UserType {
   member = 'member',
   other = 'other',
   init = 'init'
-}
-
-/** 获取当前控制台modules的 域名映射表 */
-export enum ConsoleModuleEnum {
-  /** tke-apiserver 版本 */
-  PLATFORM = 'platform',
-
-  /** 业务的版本详情 */
-  Business = 'business',
-
-  /** 通知模块 */
-  Notify = 'notify',
-
-  /** 告警模块 */
-  Monitor = 'monitor',
-
-  /** 镜像仓库 */
-  Registry = 'registry',
-
-  /** 日志模块 */
-  LogAgent = 'logagent',
-
-  /** 认证模块 */
-  Auth = 'auth',
-
-  /** 审计模块 */
-  Audit = 'audit',
-
-  /** Helm应用模块 */
-  Application = 'application'
 }
 
 export enum PlatformTypeEnum {
